requirements.txt
Babel==0.9.6
Flask==0.9
Flask-Babel==0.8
Flask-Cache==0.4.0
Flask-Mail==0.7.1
Flask-PyMongo==0.1.2
Flask-Redis==0.2
Jinja2==2.6
PyYAML==3.10
Werkzeug==0.8.3
amqplib==1.0.2
anyjson==0.3.3
billiard==2.7.3.12
blinker==1.2
celery==3.0.7
kombu==2.4.3
pymongo==2.2.1
python-dateutil==1.5
pytz==2012d
redis==2.6.0
requests==0.13.5
speaklater==1.3
wsgiref==0.1.2
